Fix weird dialog shown during login from GDM
When the session manager enters the query end session phase, it
temporarily adds all clients to the "query clients" list. This list is used to track
the progression of clients quitting. If a client quits during the query end
session phase instead of waiting until the end session phase, then it will get
destroyed but not get removed from the query clients list. This commit
removes it from the query clients list in this case. It does this by explicitly
invoking the end session response handler (in effect, the client
responded to the end session query by exiting).

This prevents a bizarre bug people are seeing where they have to click
"log out" in a dialog during the login process.

https://bugzilla.gnome.org/show_bug.cgi?id=634244