consolekit: prompt for password first before starting reboot/shutdown
Sometimes when user requests to reboot or shutdown the machine, we
need to prompt for a password (say if another user is already logged
in). Unfortunately we do this at the end of logout process, so if the
user clicks cancel they still get logged out.

This commit moves the password prompting to the front end of the
process, so a cancel cancels both the shutdown and implicit logout
operation.